# pytorrent/openapi/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "error": {
            "type": "string"
          },
          "ok": {
            "enum": [
              false
            ],
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "error"
        ],
        "type": "object"
      },
      "ApiOk": {
        "properties": {
          "ok": {
            "enum": [
              true
            ],
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "type": "object"
      },
...
      "Label": {
        "additionalProperties": true,
        "properties": {
          "color": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "Preferences": {
        "additionalProperties": true,
        "properties": {
          "bootstrap_theme": {
            "enum": [
              "default",
              "flatly",
              "litera",
              "lumen",
              "minty",
              "sketchy",
              "solar",
              "spacelab",
              "united",
              "zephyr"
            ],
            "type": "string"
          },
          "disk_monitor_mode": {
            "type": "string"
          },
          "disk_monitor_paths_json": {
            "type": "string"
          },
          "disk_monitor_selected_path": {
            "type": "string"
          },
          "font_family": {
            "enum": [
              "default",
              "system-ui",
              "figtree",
              "inter",
              "geist",
              "manrope",
              "dm-sans",
              "source-sans-3",
              "open-sans",
              "roboto",
              "lato",
              "nunito-sans",
              "poppins",
              "montserrat",
              "ibm-plex-sans",
              "jetbrains-mono",
              "adwaita-mono"
            ],
            "type": "string"
          },
          "peers_refresh_seconds": {
            "enum": [
              0,
              10,
              15,
              30,
              60
            ],
            "type": "integer"
          },
          "port_check_enabled": {
            "type": "boolean"
          },
          "reverse_dns_enabled": {
            "description": "Enables cached reverse DNS lookups for the Peers tab.",
            "type": "boolean"
          },
          "table_columns_json": {
            "description": "JSON-encoded TableColumnsPreference stored in user preferences.",
            "type": "string"
          },
          "theme": {
            "enum": [
              "light",
              "dark"
            ],
            "type": "string"
          },
          "tracker_favicons_enabled": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "Profile": {
        "additionalProperties": true,
        "properties": {
          "id": {
            "type": "integer"
          },
          "is_remote": {
            "type": "boolean"
          },
          "max_parallel_jobs": {
            "default": 5,
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "scgi_url": {
            "type": "string"
          },
          "timeout_seconds": {
            "type": "integer"
          }
        },
        "type": "object"
      },
      "ProfileInput": {
        "properties": {
          "is_remote": {
            "description": "When true, CPU/RAM host usage is hidden; public IP checks try remote rTorrent commands when supported.",
            "type": "boolean"
          },
          "max_parallel_jobs": {
            "default": 5,
            "description": "Maximum queued jobs that may run at once for this rTorrent. Move/remove jobs keep request order.",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "scgi_url": {
            "type": "string"
          },
          "timeout_seconds": {
            "type": "integer"
          }
        },
        "type": "object"
      },
      "RatioGroup": {
        "additionalProperties": true,
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "apiKeyAuth": {
        "description": "Optional per-user API token. Tokens are accepted only when PYTORRENT_AUTH_ENABLE=1. When auth is disabled, API endpoints do not require credentials.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "bearerApiToken": {
        "description": "Optional per-user API token in Authorization: Bearer <token>. Tokens are accepted only when PYTORRENT_AUTH_ENABLE=1.",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "description": "Optional Flask session cookie. Required only when PYTORRENT_AUTH_ENABLE=1 and no API token is used.",
        "in": "cookie",
        "name": "session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "pyTorrent API",
    "version": "0.0.1"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/labels": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "labels": {
                          "items": {
                            "$ref": "#/components/schemas/Label"
                          },
                          "type": "array"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "List labels"
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Label"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "labels": {
                          "items": {
                            "$ref": "#/components/schemas/Label"
                          },
                          "type": "array"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Create or update label"
      }
    },
    "/api/labels/{label_id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "label_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "labels": {
                          "items": {
                            "$ref": "#/components/schemas/Label"
                          },
                          "type": "array"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Delete label"
      }
    },
...
    "/api/preferences": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "preferences": {
                          "$ref": "#/components/schemas/Preferences"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Get preferences"
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Preferences"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "preferences": {
                          "$ref": "#/components/schemas/Preferences"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Save preferences"
      }
    },
    "/api/preferences/table-columns/recommended": {
      "post": {
        "description": "Applies the backend-owned recommended torrent table layout, including desktop visibility, mobile visibility, smart mobile filters and column widths.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "preferences": {
                          "$ref": "#/components/schemas/Preferences"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Apply recommended table columns"
      }
    },
    "/api/profiles": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "active_profile": {
                          "$ref": "#/components/schemas/Profile"
                        },
                        "profiles": {
                          "items": {
                            "$ref": "#/components/schemas/Profile"
                          },
                          "type": "array"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "List rTorrent profiles"
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "profile": {
                          "$ref": "#/components/schemas/Profile"
                        },
                        "profiles": {
                          "items": {
                            "$ref": "#/components/schemas/Profile"
                          },
                          "type": "array"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Create rTorrent profile"
      }
    },
...
    "/api/profiles/import": {
      "post": {
        "description": "Import rTorrent profiles from exported JSON.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "profiles": {
                    "items": {
                      "type": "object"
                    },
                    "type": "array"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Import profiles"
      }
    },
    "/api/profiles/test": {
      "post": {
        "description": "Test a submitted SCGI profile payload without saving it.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "is_remote": {
                    "type": "boolean"
                  },
                  "max_parallel_jobs": {
                    "type": "integer"
                  },
                  "name": {
                    "type": "string"
                  },
                  "scgi_url": {
                    "type": "string"
                  },
                  "timeout_seconds": {
                    "type": "number"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Test rTorrent profile"
      }
    },
    "/api/profiles/{profile_id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "profiles": {
                          "items": {
                            "$ref": "#/components/schemas/Profile"
                          },
                          "type": "array"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Delete rTorrent profile"
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "profile": {
                          "$ref": "#/components/schemas/Profile"
                        },
                        "profiles": {
                          "items": {
                            "$ref": "#/components/schemas/Profile"
                          },
                          "type": "array"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Update rTorrent profile"
      }
    },
    "/api/profiles/{profile_id}/activate": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "active_profile": {
                          "$ref": "#/components/schemas/Profile"
                        },
                        "profiles": {
                          "items": {
                            "$ref": "#/components/schemas/Profile"
                          },
                          "type": "array"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Activate profile"
      }
    },
...
    "/api/ratio-groups": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "groups": {
                          "items": {
                            "$ref": "#/components/schemas/RatioGroup"
                          },
                          "type": "array"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "List ratio groups"
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RatioGroup"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "groups": {
                          "items": {
                            "$ref": "#/components/schemas/RatioGroup"
                          },
                          "type": "array"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Create or update ratio group"
      }
    },
    "/api/ratio-groups/check": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Run ratio groups check"
      }
    },
...
  }
}